MPD client
- added glib-based MPD client wrapper
