+++ src/polyhymnia-application.c
@@ -5,13 +5,9 @@
 #include "polyhymnia-preferences-window.h"
 #include "polyhymnia-window.h"
 
-#include <mpd/connection.h>
-
 struct _PolyhymniaApplication
 {
   AdwApplication parent_instance;
-
-  struct mpd_connection   *mpd_connection;
 };
 
 G_DEFINE_TYPE (PolyhymniaApplication, polyhymnia_application, ADW_TYPE_APPLICATION)
@@ -51,10 +47,6 @@ polyhymnia_application_shutdown (GApplication *app)
   g_assert (POLYHYMNIA_IS_APPLICATION (app));
 
   PolyhymniaApplication *self = POLYHYMNIA_APPLICATION (app);
-  if (self->mpd_connection != NULL)
-  {
-    mpd_connection_free(self->mpd_connection);
-  }
 
   G_APPLICATION_CLASS (polyhymnia_application_parent_class)->shutdown (app);
 }
@@ -66,21 +58,6 @@ polyhymnia_application_startup (GApplication *app)
   g_assert (POLYHYMNIA_IS_APPLICATION (app));
 
   PolyhymniaApplication *self = POLYHYMNIA_APPLICATION (app);
-  self->mpd_connection = mpd_connection_new(NULL, 0, 0);
-
-  if (self->mpd_connection != NULL)
-  {
-    if (mpd_connection_get_error(self->mpd_connection) == MPD_ERROR_SUCCESS)
-    {
-      const unsigned *mpd_version = mpd_connection_get_server_version (self->mpd_connection);
-      g_debug("Connected to MPD %d.%d.%d", mpd_version[0], mpd_version[1], mpd_version[2]);
-    }
-    else
-    {
-      g_warning("An error occurred on MPD connection initialization: %s\n",
-              mpd_connection_get_error_message(self->mpd_connection));
-    }
-  }
 }
 
 static void
