+++ src/polyhymnia-mpd-client.h
@@ -1,0 +1,12 @@
+
+#pragma once
+
+#include <glib-object.h>
+
+G_BEGIN_DECLS
+
+#define POLYHYMNIA_TYPE_MPD_CLIENT (polyhymnia_mpd_client_get_type())
+
+G_DECLARE_FINAL_TYPE (PolyhymniaMpdClient, polyhymnia_mpd_client, POLYHYMNIA, MPD_CLIENT, GObject)
+
+G_END_DECLS
